requirements.txt
boto3==1.35.76
botocore==1.35.76
certifi==2024.8.30
charset-normalizer==3.4.0
fastapi==0.115.6
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-openai==0.2.11
pydantic==2.10.3
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.32.1
langgraph
python-multipart
python-dotenv
google-search-results